requirements.txt
numpy==1.26.4
onnxruntime
pillow
pydantic